requirements.txt
astroid==2.3.3
azure-core==1.1.1
azure-functions==1.0.7
azure-storage-blob==12.1.0
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
idna==2.8
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msrest==0.6.10
oauthlib==3.1.0
pycparser==2.19
pylint==2.4.4
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
soupsieve==1.9.5
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
yapf==0.29.0
